Extract SensorReceiverBase update logic into SensorUpdater Update test
