+++ common/src/main/kotlin/io/homeassistant/companion/android/common/sensors/SensorReceiverBase.kt
@@ -1,44 +1,19 @@
 package io.homeassistant.companion.android.common.sensors
 
-import android.app.NotificationChannel
-import android.app.NotificationManager
-import android.app.PendingIntent
 import android.content.BroadcastReceiver
 import android.content.Context
 import android.content.Intent
 import android.content.IntentFilter
-import android.content.res.Configuration
-import android.os.Build
-import androidx.core.app.NotificationCompat
 import androidx.core.content.ContextCompat
-import androidx.core.content.getSystemService
-import io.homeassistant.companion.android.common.R
-import io.homeassistant.companion.android.common.data.integration.IntegrationException
-import io.homeassistant.companion.android.common.data.integration.IntegrationRepository
-import io.homeassistant.companion.android.common.data.integration.SensorRegistration
 import io.homeassistant.companion.android.common.data.servers.ServerManager
-import io.homeassistant.companion.android.common.data.websocket.impl.entities.GetConfigResponse
-import io.homeassistant.companion.android.common.sensors.BluetoothSensorManager.Companion.enableDisableBeaconMonitor
-import io.homeassistant.companion.android.common.util.CHANNEL_SENSOR_SYNC
-import io.homeassistant.companion.android.common.util.SdkVersion
-import io.homeassistant.companion.android.database.sensor.SensorWithAttributes
-import io.homeassistant.companion.android.database.sensor.toSensorWithAttributes
-import io.homeassistant.companion.android.database.sensor.toSensorsWithAttributes
-import io.homeassistant.companion.android.database.server.Server
 import io.homeassistant.companion.android.database.settings.SensorUpdateFrequencySetting
 import io.homeassistant.companion.android.database.settings.SettingsDao
-import java.io.IOException
-import java.net.ConnectException
-import java.net.SocketTimeoutException
-import java.util.Locale
 import javax.inject.Inject
 import kotlin.time.Duration.Companion.seconds
 import kotlinx.coroutines.CancellationException
 import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.Job
-import kotlinx.coroutines.async
-import kotlinx.coroutines.awaitAll
 import kotlinx.coroutines.delay
 import kotlinx.coroutines.launch
 import timber.log.Timber
@@ -47,7 +22,6 @@ abstract class SensorReceiverBase : BroadcastReceiver() {
     companion object {
         const val ACTION_UPDATE_SENSOR = "io.homeassistant.companion.android.UPDATE_SENSOR"
         const val ACTION_UPDATE_SENSORS = "io.homeassistant.companion.android.UPDATE_SENSORS"
-        const val ACTION_STOP_BEACON_SCANNING = "io.homeassistant.companion.android.STOP_BEACON_SCANNING"
         const val EXTRA_SENSOR_ID = "sensorId"
 
         suspend fun SettingsDao.shouldDoFastUpdates(context: Context): Boolean {
@@ -72,9 +46,6 @@ abstract class SensorReceiverBase : BroadcastReceiver() {
 
     private val ioScope: CoroutineScope = CoroutineScope(Dispatchers.IO + Job())
 
-    protected abstract val currentAppVersion: String
-    protected abstract val managers: Set<SensorManager>
-
     @Inject
     lateinit var serverManager: ServerManager
 
@@ -87,6 +58,9 @@ abstract class SensorReceiverBase : BroadcastReceiver() {
     @Inject
     lateinit var settingsDao: SettingsDao
 
+    @Inject
+    lateinit var sensorUpdater: SensorUpdater
+
     private val chargingActions = listOf(
         Intent.ACTION_BATTERY_LOW,
         Intent.ACTION_BATTERY_OKAY,
@@ -95,13 +69,6 @@ abstract class SensorReceiverBase : BroadcastReceiver() {
     )
 
     protected abstract val skippableActions: Map<String, List<String>>
-
-    protected abstract fun getSensorSettingsIntent(
-        context: Context,
-        sensorId: String,
-        sensorManagerId: String,
-        notificationId: Int,
-    ): PendingIntent?
 
     override fun onReceive(context: Context, intent: Intent) {
         Timber.d("Received intent: ${intent.action}")
@@ -120,11 +87,6 @@ abstract class SensorReceiverBase : BroadcastReceiver() {
                     )
                     return@launch
                 }
-            }
-
-            if (intent.action == ACTION_STOP_BEACON_SCANNING) {
-                sensorRepository.enableDisableBeaconMonitor(context, false)
-                return@launch
             }
 
             @Suppress("DEPRECATION")
@@ -163,335 +125,25 @@ abstract class SensorReceiverBase : BroadcastReceiver() {
             if (intent.action == ACTION_UPDATE_SENSOR) {
                 val sensorId = intent.getStringExtra(EXTRA_SENSOR_ID)
                 if (sensorId != null) {
-                    updateSensor(context, sensorId)
+                    sensorUpdater.updateSensor(sensorId)
                 }
             } else {
-                updateSensors(context, serverManager, sensorRepository, intent)
+                updateSensors(intent)
                 if (chargingActions.contains(intent.action)) {
                     // Add a 5 second delay to perform another update so charging state updates completely.
                     // This is necessary as the system needs a few seconds to verify the charger.
                     delay(5.seconds)
-                    updateSensors(context, serverManager, sensorRepository, intent)
+                    updateSensors(intent)
                 }
             }
         }
+    }
+
+    private suspend fun updateSensors(intent: Intent?) {
+        sensorUpdater.updateSensors(intent = intent)
     }
 
     private suspend fun isSensorEnabled(id: String): Boolean {
         return sensorRepository.get(id).any { it.enabled }
-    }
-
-    suspend fun updateSensors(
-        context: Context,
-        serverManager: ServerManager,
-        sensorRepository: SensorRepository,
-        intent: Intent?,
-    ) {
-        if (!serverManager.isRegistered()) {
-            Timber.w("Device not registered, skipping sensor update/registration")
-            return
-        }
-
-        managers.forEach { manager ->
-            val hasSensor = manager.hasSensor()
-            if (hasSensor) {
-                try {
-                    manager.requestSensorUpdate(intent)
-                } catch (e: CancellationException) {
-                    throw e
-                } catch (e: Exception) {
-                    Timber.e(e, "Issue requesting updates for ${context.getString(manager.name)}")
-                }
-            }
-        }
-
-        try {
-            serverManager.servers().map { server ->
-                ioScope.async { syncSensorsWithServer(context, serverManager, server, sensorRepository) }
-            }.awaitAll()
-            Timber.i("Sensor updates and sync completed")
-        } catch (e: CancellationException) {
-            throw e
-        } catch (e: Exception) {
-            Timber.e(e, "Exception while awaiting sensor updates.")
-        }
-    }
-
-    private suspend fun syncSensorsWithServer(
-        context: Context,
-        serverManager: ServerManager,
-        server: Server,
-        sensorRepository: SensorRepository,
-    ): Boolean {
-        val config: GetConfigResponse
-        val integrationRepository: IntegrationRepository
-
-        try {
-            integrationRepository = serverManager.integrationRepository(server.id)
-            config = integrationRepository.getConfig()
-        } catch (e: CancellationException) {
-            throw e
-        } catch (e: Exception) {
-            Timber.e(e, "Error while getting core config to sync sensor status aborting")
-            return false
-        }
-
-        val currentHAversion = integrationRepository.getHomeAssistantVersion()
-        val supportsDisabledSensors = integrationRepository.isHomeAssistantVersionAtLeast(2022, 6, 0)
-        val serverIsTrusted = integrationRepository.isTrusted()
-        val coreSensorStatus: Map<String, Boolean>? =
-            if (supportsDisabledSensors && (serverIsTrusted || sensorRepository.getEnabledCount() > 0)) {
-                config.entities
-                    ?.filter { it.value["disabled"] != null }
-                    ?.mapValues { !(it.value["disabled"] as Boolean) } // Map to sensor id -> enabled
-            } else {
-                // Cannot sync disabled, or all sensors disabled and server changes aren't trusted
-                null
-            }
-
-        var serverIsReachable = true
-        val enabledRegistrations = mutableListOf<SensorRegistration<Any>>()
-
-        managers.forEach { manager ->
-            // Each manager was already asked to update in updateSensors
-            val hasSensor = manager.hasSensor()
-
-            manager.getAvailableSensors().forEach sensorForEach@{ basicSensor ->
-                val fullSensor = sensorRepository.getFull(basicSensor.id, server.id).toSensorWithAttributes()
-                val sensor = fullSensor?.sensor ?: return@sensorForEach
-                val sensorCoreEnabled = coreSensorStatus?.get(basicSensor.id)
-                val canBeRegistered = hasSensor &&
-                    basicSensor.type.isNotBlank() &&
-                    basicSensor.statelessIcon.isNotBlank()
-                val hasPermission = manager.checkPermission(basicSensor.id)
-
-                // A sensor enabled in the database but missing its runtime permission isn't really
-                // enabled. Persist that so the reconciliation below unregisters it on the server instead
-                // of leaving a stale entity behind.
-                if (sensor.enabled && !hasPermission) {
-                    sensor.enabled = false
-                    sensorRepository.update(sensor)
-                }
-
-                // Register sensor and/or update the sensor enabled state. Priority is:
-                // 1. There is a new sensor or change in enabled state according to the app
-                // 2. There is a change in enabled state according to core (user changed in frontend)
-                // 3. There is no change in enabled state, but app/core version has changed
-                if (
-                    canBeRegistered &&
-                    (
-                        (sensor.registered == null && (sensor.enabled || supportsDisabledSensors)) ||
-                            (sensor.enabled != sensor.registered && supportsDisabledSensors) ||
-                            (sensor.registered != null && coreSensorStatus != null && sensorCoreEnabled == null)
-                        )
-                ) {
-                    // 1. (Re-)register sensors with core when they can be registered and:
-                    // - sensor isn't registered, but is enabled or on core >=2022.6
-                    // - sensor enabled has changed from registered enabled state on core >=2022.6
-                    // - sensor is registered according to database, but core >=2022.6 doesn't know about it
-                    try {
-                        registerSensor(context, serverManager, fullSensor, basicSensor)
-                        sensor.registered = sensor.enabled
-                        sensor.coreRegistration = currentHAversion
-                        sensor.appRegistration = currentAppVersion
-                        sensorRepository.update(sensor)
-                    } catch (e: Exception) {
-                        Timber.e(e, "Issue registering sensor ${basicSensor.id}")
-                    }
-                } else if (
-                    canBeRegistered &&
-                    supportsDisabledSensors &&
-                    sensorCoreEnabled != null &&
-                    sensorCoreEnabled != sensor.registered
-                ) {
-                    // 2. Try updating the sensor enabled state to match core state when it's different from
-                    // the app, if the sensor can be registered, on core >= 2022.6 and server trusted.
-                    // If the server isn't trusted, update registered state to match app.
-                    try {
-                        if (!serverIsTrusted) { // Core changed, but app doesn't trust server so 'override'
-                            registerSensor(context, serverManager, fullSensor, basicSensor)
-                        } else if (sensorCoreEnabled) { // App disabled, should enable
-                            if (hasPermission) {
-                                sensor.enabled = true
-                                sensor.registered = true
-                            } else {
-                                // Can't enable due to missing permission(s), 'override' core and notify user
-                                registerSensor(context, serverManager, fullSensor, basicSensor)
-
-                                context.getSystemService<NotificationManager>()?.let { notificationManager ->
-                                    createNotificationChannel(context)
-                                    val notificationId = "$CHANNEL_SENSOR_SYNC-${basicSensor.id}".hashCode()
-                                    val notificationIntent =
-                                        getSensorSettingsIntent(context, basicSensor.id, manager.id(), notificationId)
-                                    val notification = NotificationCompat.Builder(context, CHANNEL_SENSOR_SYNC)
-                                        .setSmallIcon(R.drawable.ic_stat_ic_notification)
-                                        .setContentTitle(context.getString(basicSensor.name))
-                                        .setContentText(
-                                            context.getString(R.string.sensor_worker_sync_missing_permissions),
-                                        )
-                                        .setContentIntent(notificationIntent)
-                                        .setAutoCancel(true)
-                                        .build()
-                                    notificationManager.notify(notificationId, notification)
-                                }
-                            }
-                        } else { // App enabled, should disable
-                            sensor.enabled = false
-                            sensor.registered = false
-                        }
-
-                        sensor.coreRegistration = currentHAversion
-                        sensor.appRegistration = currentAppVersion
-                        sensorRepository.update(sensor)
-                    } catch (e: CancellationException) {
-                        throw e
-                    } catch (e: Exception) {
-                        Timber.e(e, "Issue enabling/disabling sensor ${basicSensor.id}")
-                    }
-                } else if (
-                    canBeRegistered &&
-                    serverIsReachable &&
-                    (sensor.enabled || supportsDisabledSensors) &&
-                    (currentAppVersion != sensor.appRegistration || currentHAversion != sensor.coreRegistration)
-                ) {
-                    // 3. Re-register sensors with core when they can be registered and are enabled or on
-                    // core >= 2022.6, and app or core version change is detected
-                    try {
-                        registerSensor(context, serverManager, fullSensor, basicSensor)
-                        sensor.registered = sensor.enabled
-                        sensor.coreRegistration = currentHAversion
-                        sensor.appRegistration = currentAppVersion
-                        sensorRepository.update(sensor)
-                    } catch (e: CancellationException) {
-                        throw e
-                    } catch (e: Exception) {
-                        Timber.e(e, "Issue re-registering sensor ${basicSensor.id}")
-                        if (e is IntegrationException &&
-                            (e.cause is ConnectException || e.cause is SocketTimeoutException)
-                        ) {
-                            Timber.w(
-                                "Server can't be reached, skipping other registrations for sensors due to version change",
-                            )
-                            serverIsReachable = false
-                        }
-                    }
-                }
-
-                if (canBeRegistered &&
-                    sensor.enabled &&
-                    sensor.registered != null &&
-                    (sensor.state != sensor.lastSentState || sensor.icon != sensor.lastSentIcon)
-                ) {
-                    enabledRegistrations.add(fullSensor.toSensorRegistration(basicSensor))
-                }
-            }
-        }
-
-        var success = true
-        if (enabledRegistrations.isNotEmpty()) {
-            success = try {
-                val serverSuccess = integrationRepository.updateSensors(enabledRegistrations)
-                enabledRegistrations.forEach {
-                    sensorRepository.updateLastSentStateAndIcon(it.uniqueId, it.serverId, it.state.toString(), it.icon)
-                }
-                serverSuccess
-            } catch (e: CancellationException) {
-                throw e
-            } catch (e: Exception) {
-                // Don't trigger re-registration when the server is down or job was cancelled
-                val exceptionOk = e is IntegrationException &&
-                    (e.cause is IOException)
-                if (exceptionOk) {
-                    Timber.w(e, "Exception while updating sensors")
-                } else {
-                    Timber.e(e, "Exception while updating sensors.")
-                }
-                exceptionOk
-            }
-
-            // We failed to update a sensor, we should re register next time
-            if (!success) {
-                enabledRegistrations.forEach {
-                    val sensor = sensorRepository.get(it.uniqueId, it.serverId)
-                    if (sensor != null) {
-                        sensor.registered = null
-                        sensor.lastSentState = null
-                        sensor.lastSentIcon = null
-                        sensorRepository.update(sensor)
-                    }
-                }
-            }
-        } else {
-            Timber.d("Nothing to update for server ${server.id} (${server.friendlyName})")
-        }
-        return success
-    }
-
-    private suspend fun registerSensor(
-        context: Context,
-        serverManager: ServerManager,
-        fullSensor: SensorWithAttributes,
-        basicSensor: SensorManager.BasicSensor,
-    ) {
-        val reg = fullSensor.toSensorRegistration(basicSensor)
-        val config = Configuration(context.resources.configuration)
-        config.setLocale(Locale("en"))
-        reg.name = context.createConfigurationContext(config).resources.getString(basicSensor.name)
-
-        serverManager.integrationRepository(fullSensor.sensor.serverId).registerSensor(reg)
-    }
-
-    private suspend fun updateSensor(context: Context, sensorId: String) {
-        val sensorManager = managers.firstOrNull {
-            it.getAvailableSensors().any { s -> s.id == sensorId }
-        } ?: return
-        try {
-            sensorManager.requestSensorUpdate()
-        } catch (e: CancellationException) {
-            throw e
-        } catch (e: Exception) {
-            Timber.e(e, "Issue requesting updates for ${context.getString(sensorManager.name)}")
-        }
-        val basicSensor = sensorManager.getAvailableSensors().firstOrNull { it.id == sensorId } ?: return
-        val fullSensors = sensorRepository.getFull(sensorId).toSensorsWithAttributes()
-        fullSensors.filter {
-            it.sensor.enabled &&
-                it.sensor.registered == true &&
-                (it.sensor.state != it.sensor.lastSentState || it.sensor.icon != it.sensor.lastSentIcon)
-        }.forEach { fullSensor ->
-            ioScope.launch {
-                try {
-                    serverManager.integrationRepository(
-                        fullSensor.sensor.serverId,
-                    ).updateSensors(listOf(fullSensor.toSensorRegistration(basicSensor)))
-                    sensorRepository.updateLastSentStateAndIcon(
-                        basicSensor.id,
-                        fullSensor.sensor.serverId,
-                        fullSensor.sensor.state,
-                        fullSensor.sensor.icon,
-                    )
-                } catch (e: CancellationException) {
-                    throw e
-                } catch (e: Exception) {
-                    Timber.e(e, "Exception while updating individual sensor.")
-                }
-            }
-        }
-    }
-
-    private fun createNotificationChannel(context: Context) {
-        if (SdkVersion.isAtLeast(Build.VERSION_CODES.O)) {
-            val notificationManager = context.getSystemService<NotificationManager>() ?: return
-            var notificationChannel =
-                notificationManager.getNotificationChannel(CHANNEL_SENSOR_SYNC)
-            if (notificationChannel == null) {
-                notificationChannel = NotificationChannel(
-                    CHANNEL_SENSOR_SYNC,
-                    CHANNEL_SENSOR_SYNC,
-                    NotificationManager.IMPORTANCE_DEFAULT,
-                )
-                notificationManager.createNotificationChannel(notificationChannel)
-            }
-        }
     }
 }
