+++ wear/src/main/kotlin/io/homeassistant/companion/android/sensors/SensorReceiver.kt
@@ -2,7 +2,6 @@ package io.homeassistant.companion.android.sensors
 
 import android.annotation.SuppressLint
 import android.app.NotificationManager
-import android.app.PendingIntent
 import android.bluetooth.BluetoothAdapter
 import android.content.Context
 import android.content.Intent
@@ -10,10 +9,7 @@ import android.media.AudioManager
 import android.net.wifi.WifiManager
 import android.nfc.NfcAdapter
 import android.os.PowerManager
-import androidx.core.app.TaskStackBuilder
-import androidx.core.net.toUri
 import dagger.hilt.android.AndroidEntryPoint
-import io.homeassistant.companion.android.BuildConfig
 import io.homeassistant.companion.android.common.sensors.AudioSensorManager
 import io.homeassistant.companion.android.common.sensors.BluetoothSensorManager
 import io.homeassistant.companion.android.common.sensors.DNDSensorManager
@@ -21,23 +17,10 @@ import io.homeassistant.companion.android.common.sensors.NetworkSensorManager
 import io.homeassistant.companion.android.common.sensors.NextAlarmManager
 import io.homeassistant.companion.android.common.sensors.NfcSensorManager
 import io.homeassistant.companion.android.common.sensors.PowerSensorManager
-import io.homeassistant.companion.android.common.sensors.SensorManager
 import io.homeassistant.companion.android.common.sensors.SensorReceiverBase
-import io.homeassistant.companion.android.home.HomeActivity
-import io.homeassistant.companion.android.home.views.DEEPLINK_SENSOR_MANAGER
-import javax.inject.Inject
 
 @AndroidEntryPoint
 class SensorReceiver : SensorReceiverBase() {
-
-    override val currentAppVersion: String
-        get() = BuildConfig.VERSION_NAME
-
-    @Inject
-    lateinit var injectedManagers: Set<@JvmSuppressWildcards SensorManager>
-
-    override val managers: Set<SensorManager>
-        get() = injectedManagers
 
     companion object {
         fun updateAllSensors(context: Context) {
@@ -78,22 +61,4 @@ class SensorReceiver : SensorReceiverBase() {
         BluetoothAdapter.ACTION_STATE_CHANGED to listOf(BluetoothSensorManager.bluetoothState.id),
         NfcAdapter.ACTION_ADAPTER_STATE_CHANGED to listOf(NfcSensorManager.nfcStateSensor.id),
     )
-
-    override fun getSensorSettingsIntent(
-        context: Context,
-        sensorId: String,
-        sensorManagerId: String,
-        notificationId: Int,
-    ): PendingIntent? {
-        val intent = Intent(
-            Intent.ACTION_VIEW,
-            "$DEEPLINK_SENSOR_MANAGER/$sensorManagerId".toUri(),
-            context,
-            HomeActivity::class.java,
-        )
-        return TaskStackBuilder.create(context).run {
-            addNextIntentWithParentStack(intent)
-            getPendingIntent(notificationId, PendingIntent.FLAG_UPDATE_CURRENT)
-        }
-    }
 }
