+++ wear/src/main/kotlin/io/homeassistant/companion/android/sensors/WearSensorModule.kt
@@ -1,5 +1,9 @@
 package io.homeassistant.companion.android.sensors
 
+import android.app.PendingIntent
+import android.content.Intent
+import androidx.core.app.TaskStackBuilder
+import androidx.core.net.toUri
 import dagger.Binds
 import dagger.Module
 import dagger.Provides
@@ -8,7 +12,11 @@ import dagger.hilt.components.SingletonComponent
 import dagger.multibindings.ElementsIntoSet
 import dagger.multibindings.IntoSet
 import io.homeassistant.companion.android.common.sensors.SensorManager
+import io.homeassistant.companion.android.common.sensors.SensorSettingsIntentProvider
+import io.homeassistant.companion.android.home.HomeActivity
+import io.homeassistant.companion.android.home.views.DEEPLINK_SENSOR_MANAGER
 import io.homeassistant.companion.android.sensors.generated.GeneratedProvidesSensorWear
+import javax.inject.Singleton
 
 /**
  * Hilt bindings for the `:wear` sensor managers: each wear-specific [SensorManager] contributed into
@@ -48,5 +56,21 @@ abstract class WearSensorModule {
         @Provides
         @ElementsIntoSet
         fun wearProvidesSensors(): Set<SensorManager.BasicSensor> = GeneratedProvidesSensorWear.sensors
+
+        @Provides
+        @Singleton
+        fun providesSensorSettingsIntentProvider(): SensorSettingsIntentProvider =
+            SensorSettingsIntentProvider { context, _, sensorManagerId, notificationId ->
+                val intent = Intent(
+                    Intent.ACTION_VIEW,
+                    "$DEEPLINK_SENSOR_MANAGER/$sensorManagerId".toUri(),
+                    context,
+                    HomeActivity::class.java,
+                )
+                TaskStackBuilder.create(context).run {
+                    addNextIntentWithParentStack(intent)
+                    getPendingIntent(notificationId, PendingIntent.FLAG_UPDATE_CURRENT)
+                }
+            }
     }
 }
